+++ src/adata/commands/__init__.py
@@ -6,3 +6,4 @@ from adata.commands.ls import list_store
 from adata.commands.create import create_store
 from adata.commands.split import split_store
 from adata.commands.concat import MERGE_CHOICES, MERGE_STRATEGIES, concat_stores
+from adata.commands.convert import convert_store